requirements.txt
gunicorn==21.2.0
Flask==2.3.3
Pillow==10.3.0
Werkzeug==2.3.7
ultralytics==8.1.25
torch==2.1.2
numpy==1.26.4

